+++ src/main/java/customizables/Build.java
@@ -1,10 +1,6 @@
 package customizables;
 
 import controllers.Master;
-import java.io.File;
-import java.util.*;
-import javax.json.*;
-import serialization.*;
 import util.StringUtil;
 
 
@@ -16,7 +12,7 @@ import util.StringUtil;
  * which are applied to Players at the beginning of a match
  * @author Matt
  */
-public class Build implements JsonSerialable{	
+public class Build{	
 	private final String name;
 	private final String className;
 	private final String[] activeNames;
@@ -40,88 +36,6 @@ public class Build implements JsonSerialable{
             passiveNames[1],
             passiveNames[2]
         );
-    }
-    
-    @Override
-    public JsonObject serializeJson(){
-        JsonObjectBuilder b = Json.createObjectBuilder();
-        b.add("type", "build");
-        b.add("name", name);
-        b.add("character class", className);
-        
-        JsonArrayBuilder acts = Json.createArrayBuilder();
-        for(String activeName : activeNames){
-            acts.add(activeName);
-        }
-        b.add("actives", acts.build());
-        
-        JsonArrayBuilder pass = Json.createArrayBuilder();
-        for(String passName : passiveNames){
-            pass.add(passName);
-        }
-        b.add("passives", pass.build());
-        
-        return b.build();
-    }
-    private static String getNameFrom(JsonObject obj){
-        JsonUtil.verify(obj, "name");
-        return obj.getString("name");
-    }
-    private static String getCharacterClassFrom(JsonObject obj){
-        JsonUtil.verify(obj, "character class");
-        return obj.getString("character class");
-    }
-    private static String[] getActivesFrom(JsonObject obj){
-        JsonUtil.verify(obj, "actives");
-        JsonArray a = obj.getJsonArray("actives");
-        int len = a.size();
-        String[] ret = new String[len];
-        for(int i = 0; i < len; i++){
-            ret[i] = a.getString(i);
-        }
-        return ret;
-    }
-    private static String[] getPassivesFrom(JsonObject obj){
-        JsonUtil.verify(obj, "passives");
-        JsonArray a = obj.getJsonArray("passives");
-        int len = a.size();
-        String[] ret = new String[len];
-        for(int i = 0; i < len; i++){
-            ret[i] = a.getString(i);
-        }
-        return ret;
-    }
-    public static Build deserializeJson(JsonObject obj){
-        String[] actives = getActivesFrom(obj);
-        String[] passives = getPassivesFrom(obj);
-        return new Build(
-            getNameFrom(obj),
-            getCharacterClassFrom(obj),
-            actives[0],
-            actives[1],
-            actives[2],
-            passives[0],
-            passives[1],
-            passives[2]
-        );
-    }
-    
-    public static void saveAllToFile(File f){
-        JsonObject[] objs = Arrays.stream(Master.getDataSet().getAllBuilds()).map((Build b)->{
-            return b.serializeJson();
-        }).toArray(size -> new JsonObject[size]);
-        JsonUtil.writeToFile(objs, f);  
-    }
-    
-    public static void loadFile(File f){
-        Build b = null;
-        for(JsonObject obj : JsonUtil.readFromFile(f)){
-            b = deserializeJson(obj);
-            if(b != null){
-                Master.getDataSet().addBuild(b);
-            }
-            //System.out.println(b.getDescription());
-        }
     }
     
 	public String getName(){
